Adding `g:vimroom_navigational_keys` option.
Fixing issue #2: If navigational keys were already mapped, VimroomToggle
would throw an error due to the `<unique>` flag on the mappings. I've
now wrapped those in a try/catch block to grab the error, and provided
an option for the mappings that you can turn off if you've already
mapped the keys to your liking.